requirements.txt
langchain-groq
faiss-cpu
crewai
pypdf2
python-dotenv
sentence-transformers
huggingface
crewai[tools]
onnxruntime==1.20
sentence-transformers
langchain-huggingface
googlesearch-python
scrapling
trafilatura[all]
cchardet
lxml
lxml_html_clean
lingua-language-detector